LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/block-view-timeline-current-time-vertical-rl.tentative-expected.txt


FAIL View timeline with container having vertical-rl layout assert_equals: Opacity with fill none at effect end time expected "1" but got "0.7"
